Handle the apache config using dh_apache2
The /etc/apache2/conf-available/mediawiki.conf was added, but not enabled
(i.e. symlinked to /etc/apache2/conf-enabled/mediawiki.conf). As a consequence
the mw-setup is not immediately available w/o manual setup steps.

This patch uses the debhelper addon dh_apache2 to manage this automatically,
so a lot of custom postinst code can be dropped.

Change-Id: I803288865978ee5b8e59560598bedab980e9b40d