Add a systemd unit to run runJobs.php as a service
Using --wait, the script will act as a daemon and just wait for new jobs
to come in and then immediately process them.

The main concern with such a setup is that runJobs.php needs to restart to
pick up code and configuration changes. We set --maxjobs=50 so it restarts
somewhat regularly, and set a timeout of 5 minutes, after which it will sleep
for 15 seconds and then restart.

This service is not enabled by default since after installing the package you
still need to manually install MediaWiki before this can be run.

Ideally we could set something like `After=mysql.service` but we don't
actually know the database backend being used. It's possible this will
be a race condition but the script will just crash and restart in 15
seconds, hopefully by then MySQL would be up.

This is inspired by the setup translatewiki.net has, and thanks to Nikerabbit
for explaining it to me.

Bug: T160035
Change-Id: Iec3ee0ab72f8caf8cbc1da005cd53b7b6d20e2c1