Add basic tests via autopkgtest
autopkgtest is the Debian CI system that is triggered on new package
uploads, and whenever any dependencies are updated.

The `lint` test runs a PHP linter and a JSON linter.

The `install-$dbtype` tests set up MediaWiki against the respective
database backend and then a few verifications that it functions
correctly.

Change-Id: Ia95fd9afc36ff8de6b1934b4f8cf9b8062a9651d